+++ lib/jsonable.dart
@@ -5,14 +5,13 @@ import 'package:jsonable/src/validator/validator.dart';
 
 export "./src/jsonable.dart" show Jsonable;
 export './src/validator/validator.dart' show Validator;
+export "./exceptions.dart";
+export "./src/validator/rules.dart" show Rules;
 
 typedef JsonableBuilder<E extends Jsonable> = E Function();
 
 abstract class JType<E> {
-  Validator _validator;
-  Validator get validator => this._validator;
-  JType({E value, Validator validator}) {
-    this._validator = validator;
+  JType({E value}) {
     if (value != null) {
       this._value = value;
     }
@@ -26,10 +25,9 @@ abstract class JType<E> {
 
 abstract class JString extends JType<String>
     implements Comparable<String>, Pattern {
-  JString({String initialValue, Validator validator})
+  JString({String initialValue})
       : super(
           value: initialValue,
-          validator: validator,
         );
 }
 
@@ -37,31 +35,40 @@ abstract class JNum extends JType<num> implements Comparable<num> {
   JNum({num initialValue, Validator validator})
       : super(
           value: initialValue,
-          validator: validator,
         );
 }
 
 abstract class JBool extends JType<bool> {
   JBool({bool initialValue, Validator validator})
-      : super(value: initialValue, validator: validator);
+      : super(
+          value: initialValue,
+        );
 }
 
 abstract class JList<E> extends JType<List<E>> implements List<E> {
   JList({List<E> initialValue, Validator validator})
-      : super(value: initialValue, validator: validator);
+      : super(
+          value: initialValue,
+        );
 }
 
 abstract class JClass<E extends Jsonable> extends JType<E> {
   JClass({Jsonable initialValue, Validator validator})
-      : super(value: initialValue, validator: validator);
+      : super(
+          value: initialValue,
+        );
 }
 
 abstract class JMap<E, R> extends JType<Map> implements Map<E, R> {
   JMap({Map initialValue, Validator validator})
-      : super(value: initialValue, validator: validator);
+      : super(
+          value: initialValue,
+        );
 }
 
 abstract class JDynamic extends JType<dynamic> {
   JDynamic({dynamic initialValue, Validator validator})
-      : super(value: initialValue, validator: validator);
+      : super(
+          value: initialValue,
+        );
 }