+++ lib/src/typing/CJstring.dart
@@ -1,8 +1,11 @@
 import "package:jsonable/jsonable.dart";
+import 'package:jsonable/src/validator/rules.dart';
 
 class CJstring extends JString {
-  CJstring(Jsonable parent, keyname, {String initialValue, Validator validator})
-      : super(initialValue: initialValue, validator: validator);
+  CJstring(Jsonable parent, keyname, {String initialValue, List<Rule> rules})
+      : super(
+          initialValue: initialValue,
+        );
 
   operator ==(other) {
     if (other is CJstring)